+++ R/worker.R
@@ -118,6 +118,8 @@ learner_train = function(learner, task, train_row_ids = NULL, test_row_ids = NUL
     learner = learner$clone()
   )
 
+  deadline_time_left = max(as.numeric(difftime(learner$deadline[["train"]], Sys.time(), units = "secs")), 0)
+
   # call train_wrapper with encapsulation
   result = encapsulate(
     learner$encapsulation["train"],
@@ -125,7 +127,7 @@ learner_train = function(learner, task, train_row_ids = NULL, test_row_ids = NUL
     .args = list(learner = learner, task = task),
     .pkgs = c(learner$packages, mlr_reflections$loaded_packages),
     .seed = NA_integer_,
-    .timeout = learner$timeout["train"],
+    .timeout = min(learner$timeout["train"], deadline_time_left),
     .compute = getOption("mlr3.mirai_encapsulation", "mlr3_encapsulation")
   )
   cond = cond_from_log(result$log)
@@ -305,13 +307,15 @@ learner_predict = function(learner, task, row_ids = NULL) {
       learner$model = marshal_model(learner$model, inplace = TRUE)
     }
 
+    deadline_time_left = max(as.numeric(difftime(learner$deadline[["predict"]], Sys.time(), units = "secs")), 0)
+
     result = encapsulate(
       learner$encapsulation["predict"],
       .f = predict_wrapper,
       .args = list(task = task, learner = learner),
       .pkgs = learner$packages,
       .seed = NA_integer_,
-      .timeout = learner$timeout["predict"],
+      .timeout = min(learner$timeout["predict"], deadline_time_left),
       .compute = getOption("mlr3.mirai_encapsulation", "mlr3_encapsulation")
     )